Refactor Transaction/Bucket API.
